requirements.txt
cycler==0.10.0
Django==4.1.6
django-simpleui==2023.3.1
Markdown==3.4.1
MarkupSafe==2.0.1
pendulum==2.1.2
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.0.2
pyquery==2.0.0
pytz==2021.1
pytzdata==2020.1
requests==2.25.1
simplejson==3.17.5
tqdm==4.64.0
yarg==0.1.9
yarl==1.8.2
zipp==3.15.0
